+++ packages/blocks/src/api/index.native.js
@@ -13,8 +13,8 @@ export {
 } from './serializer';
 export {
 	registerBlockType,
-	setUnknownTypeHandlerName,
-	getUnknownTypeHandlerName,
+	setUnregisteredTypeHandlerName,
+	getUnregisteredTypeHandlerName,
 	getBlockType,
 	getBlockTypes,
 	hasBlockSupport,