+++ packages/core-data/src/entities.js
@@ -9,6 +9,8 @@ import { upperFirst, camelCase, map, find } from 'lodash';
 import { addEntities } from './actions';
 import { apiFetch, select } from './controls';
 
+export const DEFAULT_ENTITY_KEY = 'id';
+
 export const defaultEntities = [
 	{ name: 'postType', kind: 'root', key: 'slug', baseURL: '/wp/v2/types' },
 	{ name: 'media', kind: 'root', baseURL: '/wp/v2/media', plural: 'mediaItems' },