+++ packages/data/src/components/with-select/index.js
@@ -3,7 +3,7 @@
  */
 import { Component } from '@wordpress/element';
 import isShallowEqual from '@wordpress/is-shallow-equal';
-import { remountOnPropChange, createHigherOrderComponent } from '@wordpress/compose';
+import { createHigherOrderComponent } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -44,11 +44,11 @@ const withSelect = ( mapSelectToProps ) => createHigherOrderComponent( ( Wrapped
 		);
 	}
 
-	const ComponentWithSelect = remountOnPropChange( 'registry' )( class extends Component {
+	class ComponentWithSelect extends Component {
 		constructor( props ) {
 			super( props );
 
-			this.subscribe();
+			this.subscribe( props.registry );
 
 			this.mergeProps = getNextMergeProps( props );
 		}
@@ -63,7 +63,19 @@ const withSelect = ( mapSelectToProps ) => createHigherOrderComponent( ( Wrapped
 		}
 
 		shouldComponentUpdate( nextProps, nextState ) {
-			const hasPropsChanged = ! isShallowEqual( this.props.ownProps, nextProps.ownProps );
+			// Cycle subscription if registry changes.
+			const hasRegistryChanged = nextProps.registry !== this.props.registry;
+			if ( hasRegistryChanged ) {
+				this.unsubscribe();
+				this.subscribe( nextProps.registry );
+			}
+
+			// Treat a registry change as equivalent to `ownProps`, to reflect
+			// `mergeProps` to rendered component if and only if updated.
+			const hasPropsChanged = (
+				hasRegistryChanged ||
+				! isShallowEqual( this.props.ownProps, nextProps.ownProps )
+			);
 
 			// Only render if props have changed or merge props have been updated
 			// from the store subscriber.
@@ -71,27 +83,28 @@ const withSelect = ( mapSelectToProps ) => createHigherOrderComponent( ( Wrapped
 				return false;
 			}
 
-			// If merge props change as a result of the incoming props, they
-			// should be reflected as such in the upcoming render.
 			if ( hasPropsChanged ) {
 				const nextMergeProps = getNextMergeProps( nextProps );
 				if ( ! isShallowEqual( this.mergeProps, nextMergeProps ) ) {
-					// Side effects are typically discouraged in lifecycle methods, but
-					// this component is heavily used and this is the most performant
-					// code we've found thus far.
-					// Prior efforts to use `getDerivedStateFromProps` have demonstrated
-					// miserable performance.
+					// If merge props change as a result of the incoming props,
+					// they should be reflected as such in the upcoming render.
+					// While side effects are discouraged in lifecycle methods,
+					// this component is used heavily, and prior efforts to use
+					// `getDerivedStateFromProps` had demonstrated miserable
+					// performance.
 					this.mergeProps = nextMergeProps;
 				}
+
+				// Regardless whether merge props are changing, fall through to
+				// incur the render since the component will need to receive
+				// the changed `ownProps`.
 			}
 
 			return true;
 		}
 
-		subscribe() {
-			const { subscribe } = this.props.registry;
-
-			this.unsubscribe = subscribe( () => {
+		subscribe( registry ) {
+			this.unsubscribe = registry.subscribe( () => {
 				if ( ! this.canRunSelection ) {
 					return;
 				}
@@ -118,7 +131,7 @@ const withSelect = ( mapSelectToProps ) => createHigherOrderComponent( ( Wrapped
 		render() {
 			return <WrappedComponent { ...this.props.ownProps } { ...this.mergeProps } />;
 		}
-	} );
+	}
 
 	return ( ownProps ) => (
 		<RegistryConsumer>