+++ packages/editor/src/components/rich-text/index.native.js
@@ -15,15 +15,39 @@ import { Component, RawHTML } from '@wordpress/element';
 import { withInstanceId, compose } from '@wordpress/compose';
 import { Toolbar } from '@wordpress/components';
 import {
+	isEmpty,
 	create,
 	split,
 	toHTMLString,
 } from '@wordpress/rich-text';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { FORMATTING_CONTROLS } from './formatting-controls';
+
+const FORMATTING_CONTROLS = [
+	{
+		icon: 'editor-bold',
+		title: __( 'Bold' ),
+		format: 'bold',
+	},
+	{
+		icon: 'editor-italic',
+		title: __( 'Italic' ),
+		format: 'italic',
+	},
+	{
+		icon: 'admin-links',
+		title: __( 'Link' ),
+		format: 'link',
+	},
+	{
+		icon: 'editor-strikethrough',
+		title: __( 'Strikethrough' ),
+		format: 'strikethrough',
+	},
+];
 
 const isRichTextValueEmpty = ( value ) => {
 	return ! value || ! value.length;
@@ -41,24 +65,15 @@ export class RichText extends Component {
 		super( ...arguments );
 		this.onChange = this.onChange.bind( this );
 		this.onEnter = this.onEnter.bind( this );
+		this.onBackspace = this.onBackspace.bind( this );
 		this.onContentSizeChange = this.onContentSizeChange.bind( this );
 		this.changeFormats = this.changeFormats.bind( this );
 		this.toggleFormat = this.toggleFormat.bind( this );
 		this.onActiveFormatsChange = this.onActiveFormatsChange.bind( this );
-		this.onHTMLContentWithCursor = this.onHTMLContentWithCursor.bind( this );
 		this.state = {
 			formats: {},
 			selectedNodeId: 0,
 		};
-	}
-
-	// eslint-disable-next-line no-unused-vars
-	onHTMLContentWithCursor( htmlText, start, end ) {
-		if ( ! this.props.onSplit ) {
-			// TODO: insert the \n char instead?
-			return;
-		}
-		this.splitContent( htmlText, start, end );
 	}
 
 	/*
@@ -87,10 +102,18 @@ export class RichText extends Component {
 		} );
 
 		// TODO : Fix the index position in AztecNative for Android
-		let [ before, after ] = split( { start: start - 6, end: end - 6, ...record } );
+		let [ before, after ] = split( { start, end, ...record } );
 
-		// TODO : Handle here the cases when the split happens at the trailing or leading edge...
-		// See the web version for reference.
+		// In case split occurs at the trailing or leading edge of the field,
+		// assume that the before/after values respectively reflect the current
+		// value. This also provides an opportunity for the parent component to
+		// determine whether the before/after value has changed using a trivial
+		//  strict equality operation.
+		if ( isEmpty( after ) ) {
+			before = record;
+		} else if ( isEmpty( before ) ) {
+			after = record;
+		}
 
 		if ( before ) {
 			before = this.valueToFormat( before );
@@ -100,14 +123,19 @@ export class RichText extends Component {
 			after = this.valueToFormat( after );
 		}
 
+		// The onSplit event can cause a content update event for this block.  Such event should
+		// definitely be processed by our native components, since they have no knowledge of
+		// how the split works.  Setting lastEventCount to undefined forces the native component to
+		// always update when provided with new content.
+		this.lastEventCount = undefined;
+
 		onSplit( before, after );
 	}
 
 	valueToFormat( { formats, text } ) {
 		const value = toHTMLString( { formats, text }, this.multilineTag );
-		// remove the outer p tags
-		const returningContentWithoutParaTag = value.replace( /<p>|<\/p>/gi, '' );
-		return returningContentWithoutParaTag;
+		// remove the outer root tags
+		return this.removeRootTagsProduceByAztec( value );
 	}
 
 	onActiveFormatsChange( formats ) {
@@ -124,6 +152,18 @@ export class RichText extends Component {
 			selectedNodeId: this.state.selectedNodeId + 1,
 		} );
 	}
+
+	/*
+	 * Cleans up any root tags produced by aztec.
+	 * TODO: This should be removed on a later version when aztec doesn't return the top tag of the text being edited
+	 */
+
+	removeRootTagsProduceByAztec( html ) {
+		const openingTagRegexp = RegExp( '^<' + this.props.tagName + '>', 'gim' );
+		const closingTagRegexp = RegExp( '</' + this.props.tagName + '>$', 'gim' );
+		return html.replace( openingTagRegexp, '' ).replace( closingTagRegexp, '' );
+	}
+
 	/**
 	 * Handles any case where the content of the AztecRN instance has changed.
 	 */
@@ -134,11 +174,7 @@ export class RichText extends Component {
 			clearTimeout( this.currentTimer );
 		}
 		this.lastEventCount = event.nativeEvent.eventCount;
-		// The following method just cleans up any root tags produced by aztec and replaces them with a br tag
-		// This should be removed on a later version when aztec doesn't return the top tag of the text being edited
-		const openingTagRegexp = RegExp( '^<' + this.props.tagName + '>', 'gim' );
-		const closingTagRegexp = RegExp( '</' + this.props.tagName + '>$', 'gim' );
-		const contentWithoutRootTag = event.nativeEvent.text.replace( openingTagRegexp, '' ).replace( closingTagRegexp, '' );
+		const contentWithoutRootTag = this.removeRootTagsProduceByAztec( event.nativeEvent.text );
 		this.lastContent = contentWithoutRootTag;
 		// Set a time to call the onChange prop if nothing changes in the next second
 		this.currentTimer = setTimeout( function() {
@@ -161,8 +197,23 @@ export class RichText extends Component {
 		);
 	}
 
-	onEnter() {
-		this._editor.requestHTMLWithCursor();
+	// eslint-disable-next-line no-unused-vars
+	onEnter( event ) {
+		if ( ! this.props.onSplit ) {
+			// TODO: insert the \n char instead?
+			return;
+		}
+
+		this.splitContent( event.nativeEvent.text, event.nativeEvent.selectionStart, event.nativeEvent.selectionEnd );
+	}
+
+	// eslint-disable-next-line no-unused-vars
+	onBackspace( event ) {
+		if ( ! this.props.onBackspace ) {
+			return;
+		}
+
+		this.onBackspace( event.nativeEvent.text, event.nativeEvent.selectionStart, event.nativeEvent.selectionEnd );
 	}
 
 	shouldComponentUpdate( nextProps ) {
@@ -251,9 +302,9 @@ export class RichText extends Component {
 					text={ { text: html, eventCount: this.lastEventCount } }
 					onChange={ this.onChange }
 					onEnter={ this.onEnter }
+					onBackspace={ this.onBackspace }
 					onContentSizeChange={ this.onContentSizeChange }
 					onActiveFormatsChange={ this.onActiveFormatsChange }
-					onHTMLContentWithCursor={ this.onHTMLContentWithCursor }
 					color={ 'black' }
 					maxImagesWidth={ 200 }
 					style={ style }