+++ phpunit/class-gutenberg-rest-api-test.php
@@ -381,17 +381,7 @@ class Gutenberg_REST_API_Test extends WP_Test_REST_TestCase {
 		$request = new WP_REST_Request( 'GET', '/wp/v2/users' );
 		$request->set_param( 'per_page', '-1' );
 		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 200, $response->get_status() );
-	}
-
-	public function test_get_items_unbounded_per_page_unauthorized() {
-		wp_set_current_user( $this->subscriber );
-		$request = new WP_REST_Request( 'GET', '/wp/v2/users' );
-		$request->set_param( 'per_page', '-1' );
-		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 403, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertEquals( 'rest_forbidden_per_page', $data['code'] );
+		$this->assertEquals( 400, $response->get_status() );
 	}
 
 	public function test_get_categories_unbounded_per_page() {
@@ -400,18 +390,7 @@ class Gutenberg_REST_API_Test extends WP_Test_REST_TestCase {
 		$request = new WP_REST_Request( 'GET', '/wp/v2/categories' );
 		$request->set_param( 'per_page', '-1' );
 		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 200, $response->get_status() );
-	}
-
-	public function test_get_categories_unbounded_per_page_unauthorized() {
-		wp_set_current_user( $this->subscriber );
-		$this->factory->category->create();
-		$request = new WP_REST_Request( 'GET', '/wp/v2/categories' );
-		$request->set_param( 'per_page', '-1' );
-		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 403, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertEquals( 'rest_forbidden_per_page', $data['code'] );
+		$this->assertEquals( 400, $response->get_status() );
 	}
 
 	public function test_get_pages_unbounded_per_page() {
@@ -420,18 +399,7 @@ class Gutenberg_REST_API_Test extends WP_Test_REST_TestCase {
 		$request = new WP_REST_Request( 'GET', '/wp/v2/pages' );
 		$request->set_param( 'per_page', '-1' );
 		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 200, $response->get_status() );
-	}
-
-	public function test_get_pages_unbounded_per_page_unauthorized() {
-		wp_set_current_user( $this->subscriber );
-		$this->factory->post->create( array( 'post_type' => 'page' ) );
-		$request = new WP_REST_Request( 'GET', '/wp/v2/pages' );
-		$request->set_param( 'per_page', '-1' );
-		$response = rest_get_server()->dispatch( $request );
-		$this->assertEquals( 403, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertEquals( 'rest_forbidden_per_page', $data['code'] );
+		$this->assertEquals( 400, $response->get_status() );
 	}
 
 	public function test_get_post_links_predecessor_version() {