+++ packages/editor/CHANGELOG.md
@@ -1,4 +1,29 @@
-## 5.0.0 (Unreleased)
+## 6.0.0 (unreleased)
+
+### Breaking Changes
+
+- The `labels.name` property has been removed from `MediaPlaceholder` in favor of the new `labels.instructions` prop.
+- The `UnsavedChangesWarning` component no longer accepts a `forceIsDirty` prop.
+- `mediaDetails` in object passed to `onFileChange` callback of `mediaUpload`. Please use `media_details` property instead.
+
+### New Features
+
+- In `MediaPlaceholder`, provide default values for title and instructions labels when allowed type is one of image, audio or video.
+- New actions `lockPostSaving` and `unlockPostSaving` were introduced ([#10649](https://github.com/WordPress/gutenberg/pull/10649)).
+- New selector `isPostSavingLocked` was introduced ([#10649](https://github.com/WordPress/gutenberg/pull/10649)).
+
+### Polish
+
+- Add animated logo to preview interstitial screen.
+- Tweak the editor styles support.
+
+### Bug Fixes
+
+- Made preview interstitial text translatable.
+
+## 5.0.1 (2018-10-22)
+
+## 5.0.0 (2018-10-19)
 
 ### Breaking Changes
 
@@ -7,6 +32,12 @@
 ### Deprecations
 
 - `wp.editor.PanelColor` has been deprecated in favor of `wp.editor.PanelColorSettings`.
+
+### New Features
+
+- Added `onClose` prop to `URLPopover` component.
+
+## 4.0.3 (2018-10-18)
 
 ## 4.0.0 (2018-09-30)
 
@@ -44,7 +75,7 @@
 - `isFetchingSharedBlock` selector has been removed. Use `isFetchingReusableBlock` instead.
 - `getSharedBlocks` selector has been removed. Use `getReusableBlocks` instead.
 - `editorMediaUpload` has been removed. Use `mediaUpload` instead.
-- Change how required built-ins are polyfilled with Babel 7 ([#9171](https://github.com/WordPress/gutenberg/pull/9171)).  If you're using an environment that has limited or no support for ES2015+ such as lower versions of IE then using [core-js](https://github.com/zloirock/core-js) or [@babel/polyfill](https://babeljs.io/docs/en/next/babel-polyfill) will add support for these methods.
+- Change how required built-ins are polyfilled with Babel 7 ([#9171](https://github.com/WordPress/gutenberg/pull/9171)). If you're using an environment that has limited or no support for ES2015+ such as lower versions of IE then using [core-js](https://github.com/zloirock/core-js) or [@babel/polyfill](https://babeljs.io/docs/en/next/babel-polyfill) will add support for these methods.
 - `wp.editor.DocumentTitle` component has been removed.
 - `getDocumentTitle` selector (`core/editor`) has been removed.